Move `TextSelectCopy` to `shared`
