+++ web/packages/shared/components/TextSelectCopy/TextSelectCopy.tsx
