+++ web/packages/shared/components/TextSelectCopy/index.tsx
